+++ packages/services/grpc/ecb-provider/src/consts.ts
@@ -1,1 +1,2 @@
+// eslint-disable-next-line import/prefer-default-export
 export const BASE_CURRENCY = 'EUR';
